% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_alluvial_plot_function.R
\name{alluvial_new}
\alias{alluvial_new}
\title{Alluvial plot}
\usage{
alluvial_new(..., freq, col = "gray", border = 0, layer, hide = FALSE,
  alpha = 0.5, gap.width = 0.05, xw = 0.1, cw = 0.1, blocks = TRUE,
  ordering = NULL, axis_labels = NULL, cex = par("cex"),
  cex.axis = par("cex.axis"), rotate = 0)
}
\description{
this function produces an alluvial diagram. This function is a copy of alluvial::alluvial
}

