% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_function.r
\name{forestplt}
\alias{forestplt}
\title{Forest plot for subgroup effect size}
\usage{
forestplt(dat, covari.sel, trt.sel, resp.sel, outcome.type,
  size.shape = c(0.25, 0.12), font.size = c(1.3, 1, 0.85, 0.9),
  title = NULL, lab.x = NULL)
}
\arguments{
\item{dat:}{a data set}

\item{covari.sel:}{a vector of indices of the two covariates}

\item{trt.sel:}{a covariate index specifying the treatment code}

\item{resp.sel:}{a covariate index specifying the response variable}

\item{outcome.type:}{a string specifying the type of the response variable, it can be "continuous", or "binary" or  "survival".}

\item{size.shape:}{a vector specifying the height and width of the diamonds displaying sample sizes.}

\item{font.size:}{a vector specifying the size of labels and text; the first element is for the main titles, the second is for
for the x-axis labels; the thrid is for the text in the first sub-figure; the fourth is for the unit labels of
the x-axis.}

\item{title:}{a list of three strings specifying the main titles of the three sub-figures.}

\item{lab.x:}{a list of three strings specifying the x-axis labels of the three sub-figures.}
}
\description{
This function produces a forest plot showing the treatment effect size of subgroups defined by the categories of covariates. The
first sub-figure provides a table of treatment effect estimate and sample size (for treatment / control group within each subgroup)
; the second sub-figure shows forest plots for subgroups and full population; the third displays forest plots of treatment and
control group for each population. The dashed vertical line indicates no effect Note that the overall size of diamonds which
represent subgroups can be adjusted by setting different values on the associated input argument. In addition, the function uses
log odd ratio and log hazard ratio for displaying subgroup effect sizes in binary and survival data, respectively.
}
