% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_plot_function.r
\name{interactionplt}
\alias{interactionplt}
\title{Interaction plot for subgroup effect}
\usage{
interactionplt(dat, covari.sel, trt.sel, resp.sel, outcome.type,
  add.aux.line = FALSE, font.size = c(15, 10, 0.5), title = NULL,
  lab.y = NULL)
}
\arguments{
\item{dat:}{a data set}

\item{covari.sel:}{a vector of indices of the two covariates}

\item{trt.sel:}{a covariate index specifying the treatment code}

\item{resp.sel:}{a covariate index specifying the response variable}

\item{outcome.type:}{a string specifying the type of the response variable, it can be "continuous", or "binary" or  "survival".}

\item{add.aux.line:}{a logical operator displaying the auxiliary horizontal line for checking heterogeneity in treatment effects if TRUE}

\item{font.size:}{a vector specifying the size of labels and text; the first element is for the main title and  the second element}

\item{is}{for the text in the left, right and bottom labels; the third is for the unit labels on the y-axis.}

\item{title:}{a string specifying the main title.}

\item{lab.y:}{a string specifying the y-axis label}
}
\description{
Interaction plot for subgroup effect
}
