% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-venn-diagram-function-new-fill.r
\name{plot_venn_fill}
\alias{plot_venn_fill}
\title{Venn diagram for subgroup effect size}
\usage{
plot_venn_fill(dat, covari.sel, cat.sel, trt.sel, resp.sel, outcome.type,
  range.strip = c(-6, 6), n.brk = 13, n.brk.axis = 7, font.size = c(1,
  1.5, 1, 0.9, 1, 1), title = NULL, strip = NULL, effect = "HR",
  show.overall = TRUE, palette = "divergent", col.power = 0.5)
}
\arguments{
\item{dat}{a data set}

\item{covari.sel}{a vector of indices of covariates}

\item{cat.sel}{a vector of indices of the categories for each covariate}

\item{trt.sel}{a covariate index specifying the treatment code}

\item{resp.sel}{a covariate index specifying the response variable}

\item{outcome.type}{a string specifying the type of the response variable, it can be "continuous", or "binary" or  "survival".}

\item{range.strip}{a vector with two elements specifying the range of treatment effect size for display}

\item{n.brk}{a number specifying the number of the points dividing the range of the argument "range.strip".}

\item{n.brk.axis}{a number specifying the number of breakpoints dividing the axis of the argument "range.strip".}

\item{font.size}{a vector specifying the size of labels and text; the first element is for the main title; the second is for the category labels;
the third is for the sample size labels; the fourth is for the legend text; the fifth is for the y-axis label of the colour strip;
the sixth is for the unit label on the y axis.}

\item{title}{a string specifying the main title.}

\item{strip}{a string specifying the title of the colour strip.}

\item{effect}{either "HR" or "RMST". only when outcome.type = "survival"}

\item{show.overall}{logical. whether to show or not the overall treatment effect in the strip}

\item{palette}{either "divergent" or "hcl"}

\item{col.power}{to be used when palette = "hcl". see colorspace package for reference}
}
\description{
This function produces a Venn diagram showing the treatment effect size of subgroups defined by sets from the categories of covariates.
Also, it prints out the minimum and maximum of the treatment effect size on the console so as to set an approapriate range for effect
size on the colour strip . Note that there are two options of graphical display; whether show the subgroup effect size of the complement
of the union of all the considered subgroups or not. In addition, this function only works up to 5 sets and does not run an area-proportional
algorithms for displaying two or three set. In addition, the function uses log odd ratio and log hazard ratio for displaying
subgroup effect sizes in binary and survival data, respectively.
}
\examples{
data(prca)
plot_venn_fill(prca,
        covari.sel = c(5, 7, 4),#vars,
        cat.sel = c(2,2,2),
        trt.sel = 3,
        resp.sel = c(1,2),
        outcome.type = "survival",
        range.strip = c(-3, 3),
        n.brk = 31,
        n.brk.axis = 7,
        font.size = c(0.5, 0.5, 0.7, 0.5, 0.6, 0.6),
        strip = paste("Treatment effect size (log hazard ratio)"), palette = "hcl")
}

