% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-data.R
\docType{data}
\name{prca}
\alias{prca}
\title{Prostate Cancer Dataset}
\format{A data frame with 475 rows and 15 variables:
\describe{
  \item{survtime}{survival time. Response variable}
  \item{cens}{The status indicator, 0=alive, 1=dead}
  \item{rx}{treatment received, 0=control, 1=treatment}
  \item{bm}{existence of bone metastasis}
  \item{hx}{history of cardiovascular events}
  \item{stage}{disease stage (3 or 4)}
  \item{pf}{performance}
  \item{age}{age}
  \item{weight}{weight in kg.}
  \item{age1}{dummy variable for 65 <= age < 75}
  \item{age2}{dummy variable for age >= 75}
  \item{wt1}{dummy variable for 90 <= weight < 110}
  \item{wt2}{dummy variable for weight >= 110}
  \item{age_group}{age categorized in 3 groups}
  \item{weight_group}{weight categorized in 3 groups}
}}
\source{
\url{http://portal.uni-freiburg.de/imbi/Royston-Sauerbrei-book/index.html}

\url{https://onlinelibrary.wiley.com/doi/full/10.1002/bimj.201500147}
}
\usage{
data(prca)
}
\description{
a prostate carcinoma dataset from a clinical trial David P. Byar and
Sylvan B. Green. The choice of treatment for cancer patients based on
covariate information: application to prostate cancer.
Bulletin du Cancer, 67:477–490, 1980.
The data can be found in the web:
\url{http://portal.uni-freiburg.de/imbi/Royston-Sauerbrei-book/index.html}.
We downloaded it form the supplementary material of Rosenkranz (2016)
https://onlinelibrary.wiley.com/doi/full/10.1002/bimj.201500147
 modified the data to keep relevant variables, and created categorical
 ones from age and weight.
}
\keyword{datasets}

