% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-mosaic-plot-function.R
\name{plot_mosaic}
\alias{plot_mosaic}
\title{Circular plot using circlize package}
\usage{
plot_mosaic(dat, covari.sel, trt.sel, resp.sel, outcome.type, range.v = NULL,
  adj.ann.subgrp = 4, range.strip = c(-3, 3), n.brk = 7,
  font.size = c(1, 1, 0.85, 0.85, 1), title = NULL, lab.xy = NULL,
  strip = "Treatment effect size", effect = "HR", lwd. = 2, sep. = 0.05,
  palette = "divergent", col.power = 0.5)
}
\arguments{
\item{dat:}{a data set}

\item{covari.sel:}{a vector of indices of the two covariates}

\item{trt.sel:}{a covariate index specifying the treatment code}

\item{resp.sel:}{a covariate index specifying the response variable}

\item{outcome.type:}{a string specifying the type of the response variable, it can be "continuous", or "binary" or  "survival".}

\item{range.v:}{a vector specifying the vertical range of graphical display.}

\item{adj.ann.subgrp:}{a parameter adjusting the distance between a point and its corresponding subgroup label. The smaller the value
is, the larger the distance is.}

\item{font.size:}{a vector specifying the size of labels and text; the first element is for the main title, the second is for
for x-axis and y-axis labels; the thrid is for the legend text of subgroups; the fourth is for the subgroup
labels near points; the fifth is for the unit labels on all the axes.}

\item{title:}{a string specifying the main title.}

\item{lab.xy:}{a list of two strings specifying the labels of the x and y axes.}
}
\description{
This function produces a circular plot for subgroup analysis
}

